GraVL-BERT/requirements.txt
easydict
jsonlines
matplotlib
networkx
numpy
opencv-python
pandas
Pillow
protobuf==3.10.0
pycocotools
PyYAML
regex==2019.8.19
requests==2.22.0
scikit-image
scipy
tensorboard
tensorboardX
tensorflow
tqdm
urllib3
boto3
transformers
jinja2
ninja
